Converted .tor into a cached_property
